+++ packages/core/src/session/stateless/refreshUserInfo.ts
@@ -4,8 +4,8 @@ import type { InternalStatelessContext } from "@/@types/internal.ts"
 
 export const refreshUserInfo = <DefaultUser extends User = User>({ ctx, cookies, cookieManager }: InternalStatelessContext) => {
     const refreshSession = __refreshSession<DefaultUser>({ ctx, cookies, cookieManager })
-    return async (userInfo: Partial<DefaultUser>, headers: Headers, skipCSRFCheck?: boolean) => {
-        const value = await refreshSession(headers, { user: userInfo as DefaultUser }, skipCSRFCheck)
+    return async (userInfo: Partial<DefaultUser>, headers: Headers) => {
+        const value = await refreshSession(headers, { user: userInfo as DefaultUser })
         return value
     }
 }